Outlet strana uradjena sa svim filterima

+++ src/app/components/shop/product/product.component.ts
@@ -9,7 +9,8 @@ import { ProductService } from 'src/app/services/product.service';
   styleUrls: ['./product.component.css'],
 })
 export class ProductComponent implements OnInit {
-  @Input() model: Product;
+  @Input() model;
+  @Input() typeOfProduct: number;
   @Input() adminView: boolean;
 
   @Output() productDeleted: EventEmitter<any> = new EventEmitter<any>();
@@ -17,20 +18,41 @@ export class ProductComponent implements OnInit {
   constructor(private router: Router, private productService: ProductService) {}
 
   ngOnInit(): void {
-    this.model.productDetailsFolderUrl =
-      'http://localhost/ClothesShopApi/product/product-details-images/' +
-      this.model.picUrl.substr(47, this.model.picUrl.length);
+    if (this.typeOfProduct != 1)
+      this.model.productDetailsFolderUrl =
+        'http://localhost/ClothesShopApi/product/product-details-images/' +
+        this.model.picUrl.substr(47, this.model.picUrl.length);
+    else if (this.typeOfProduct === 1)
+      this.model.productDetailsFolderUrl =
+        'http://localhost/ClothesShopApi/userProduct/product-details-images/' +
+        this.model.picUrl.substr(52, this.model.picUrl.length);
   }
 
   deleteProduct() {
-    this.productService.deleteProduct(this.model.id).subscribe(
-      (res) => {
-        console.log(res);
-        this.productDeleted.emit();
-      },
-      (err) => {
-        console.log(err);
+    if (this.adminView) {
+      if (this.typeOfProduct != 1) {
+        this.productService.deleteProduct(this.model.id).subscribe(
+          (res) => {
+            console.log(res);
+            this.productDeleted.emit();
+          },
+          (err) => {
+            console.log(err);
+          }
+        );
+      } else if (this.typeOfProduct === 1) {
+        var token = localStorage.getItem('token');
+
+        this.productService.deleteUserProduct(this.model.id, token).subscribe(
+          (res) => {
+            console.log(res);
+            this.productDeleted.emit();
+          },
+          (err) => {
+            console.log(err);
+          }
+        );
       }
-    );
+    }
   }
 }
